requirements.txt
# Core web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# OpenAI API
openai>=1.3.7

# Vector database
chromadb>=0.4.18

# Machine learning
torch>=2.1.0
transformers>=4.35.0
sentence-transformers>=2.2.2
accelerate>=0.24.1

# OpenTelemetry core
opentelemetry-api>=1.37.0
opentelemetry-sdk>=1.37.0
opentelemetry-instrumentation>=0.58b0

# OpenTelemetry instrumentations
opentelemetry-instrumentation-fastapi>=0.58b0
opentelemetry-instrumentation-requests>=0.58b0
opentelemetry-instrumentation-httpx>=0.58b0
opentelemetry-instrumentation-sqlite3>=0.58b0

# OpenTelemetry exporters
opentelemetry-exporter-otlp>=1.37.0
opentelemetry-exporter-prometheus>=0.58b0

# OpenTelemetry semantic conventions
opentelemetry-semantic-conventions>=0.58b0

# HTTP clients
httpx>=0.25.2
requests>=2.31.0

# Environment management
python-dotenv>=1.0.0

# Data processing
numpy>=1.26.0
pandas>=2.1.4

# System monitoring
psutil>=5.9.6
GPUtil>=1.4.0
pynvml>=11.5.0

# Development tools
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1

# Logging
structlog>=23.2.0

# Validation
pydantic>=2.5.0

# Metrics
prometheus-client>=0.19.0

# OpenLIT for AI observability and evaluation
openlit>=1.35.0